net: framework: Move net/family
Switch socket family declaration from embox/net/family.h to
socket/family.h
