--- src/include/embox/net/family.h
@@ -1,15 +1,0 @@
-/**
- * @file
- * @brief Main Embox netfamily include file for user netpacks.
- *
- * @date 12.05.13
- * @author Ilia Vaprol
- */
-
-#ifndef EMBOX_NET_FAMILY_H_
-#define EMBOX_NET_FAMILY_H_
-
-#include <framework/net/family/api.h>
-#include <framework/net/family/self.h>
-
-#endif /* EMBOX_FAMILY_H_ */
