+++ src/net/socket/af_inet.c
@@ -8,22 +8,25 @@
  */
 
 #include <errno.h>
-#include <framework/net/sock/api.h>
+
 #include <stddef.h>
 #include <arpa/inet.h>
-#include <net/sock.h>
 #include <sys/socket.h>
-#include <net/socket/inet_sock.h>
-#include <net/inetdevice.h>
-#include <embox/net/family.h>
 #include <string.h>
-#include <mem/misc/pool.h>
-#include <net/l3/route.h>
-#include <net/l3/ipv4/ip.h>
 #include <netinet/in.h>
 
-#include <net/socket/inet_sock.h>
+#include <net/sock.h>
+#include <net/inetdevice.h>
 
+#include "family.h"
+#include <framework/net/sock/api.h>
+
+#include <mem/misc/pool.h>
+
+#include <net/l3/route.h>
+#include <net/l3/ipv4/ip.h>
+
+#include <net/socket/inet_sock.h>
 #include <framework/mod/options.h>
 
 #define MODOPS_AMOUNT_INET_SOCK OPTION_GET(NUMBER, amount_inet_sock)
