+++ src/net/socket/af_packet.c
@@ -8,20 +8,26 @@
 
 #include <errno.h>
 #include <stdlib.h>
-#include <mem/misc/pool.h>
-#include <kernel/sched/sched_lock.h>
-#include <framework/mod/options.h>
-#include <framework/net/sock/self.h>
-#include <embox/net/family.h>
-#include <net/sock.h>
 #include <sys/socket.h>
+#include <sys/uio.h>
+
+#include <mem/misc/pool.h>
+
+
+#include <framework/net/sock/self.h>
+#include "family.h"
+
+#include <net/sock.h>
 #include <net/l3/ipv4/ip.h>
 #include <net/l2/ethernet.h>
 #include <net/if.h>
 #include <netpacket/packet.h>
+
 #include <net/sock_wait.h>
+#include <kernel/sched/sched_lock.h>
+
 #include <net/socket/packet.h>
-#include <sys/uio.h>
+#include <framework/mod/options.h>
 
 #define MODOPS_AMOUNT_SOCKETS OPTION_GET(NUMBER, amount_sockets)
